Add files via upload
Image processing - The objective of this case study is to create at Convolutional Neural Network model to predict whether or not a patient has pneumonia or not, based on x-ray images.

Text analysis - The aim of this Notebook is to use Convolutional Neural Networks or Recurrent Neural Networks in order to create a multiclass classifier for patient drug reviews from Drugs.com. The goal of the neural network model is to anticipate a drug's rating based on the content of the review. To make the process easier and to demonstrate sentiment analysis in this case study, the ratings are divided into three categories: Positive, Neutral, and Negative.